replaced SubMenu with BountyMenu and moved BountyHeading to all views of BountyPage
