+++ components/Bounty/CopyBountyAddress.js
@@ -2,10 +2,10 @@
 import React from 'react';
 import CopyAddressToClipboard from '../Copy/CopyAddressToClipboard';
 
-const CopyBountyAddress = ({ address }) => {
+const CopyBountyAddress = ({ address, styles }) => {
 
 	return (
-		<div className="">
+		<div className={styles || ''}>
 			<div className="flex  items-center space-x-2 cursor-pointer">
 				<CopyAddressToClipboard data={address} /> 
 			</div>
